Changed calculation of cross-section S
The calculation of the cross-sectional area S is not explicitly stated in the paper.
The previous equation (area of circle) described the inner cross-section of a cylindrical coil which was probably not suitable. 
The new equation describes the cross-section of the conducting coil material (height * width * fill factor). 
Therefore, height hc was added to the model parameters.
Old equation was commented out for reference .